+++ app/models/program.rb
@@ -12,6 +12,22 @@ class Program < ApplicationRecord
 
   public
 
+  def current_week
+    ((Date.current - created_at.to_date).to_i / 7) + 1
+  end
+
+  def current_routine
+    week = current_week
+    cumulative_weeks = 0
+
+    # Assuming routines are executed in order of ID
+    routines.order(:id).find do |routine|
+      duration = routine.duration_weeks || 4 # Default to 4 weeks if not specified
+      cumulative_weeks += duration
+      week <= cumulative_weeks
+    end || routines.last
+  end
+
   def assign_to_user(target_user)
     transaction do
       new_program = dup